Fixes for 64-bit Windows
git-svn-id: https://dslabpc10.epfl.ch/svn/src/s2e/trunk@9534 f94e2f5b-dfd8-4992-a279-46519bc7c46e

+++ klee/lib/Core/AddressSpace.cpp
@@ -309,7 +309,7 @@ void AddressSpace::copyOutConcretes() {
     assert(!mo->isUserSpecified);
 
     const ObjectState *os = it->second;
-    uint8_t *address = (uint8_t*) (unsigned long) mo->address;
+    uint8_t *address = (uint8_t*) (uintptr_t) mo->address;
 
     if (!os->readOnly)
       memcpy(address, os->concreteStore, mo->size);
@@ -323,7 +323,7 @@ bool AddressSpace::copyInConcretes() {
     assert(!mo->isUserSpecified);
 
     const ObjectState *os = it->second;
-    uint8_t *address = (uint8_t*) (unsigned long) mo->address;
+    uint8_t *address = (uint8_t*) (uintptr_t) mo->address;
 
     if (os->readOnly) {
       if (memcmp(address, os->concreteStore, mo->size)!=0)
